Update tests to account for storage of the initial state
